Remove the Debian installer's Java 7 dependency.
More specifically, remove the alternative dependency from the Debian
installer's control file. This was originally added back when Java 6 was
still in common use to handle cases where the default JRE was too old.
However, with the obsolescence of Java 7, it is no longer needed, and users
are now reporting incorrect notifications about it. To simplify the
installation process, remove it.

This will fix #22.